requirements.txt
streamlit
google-generativeai
pandas
streamlit-lottie==0.0.5